Changes needed to get the panel to use ORBit.
Changes needed to get the panel to use ORBit.

Please let me know if this breaks compilation with MICO - I'm not fond of
MICO but I don't want to destabilize the panel by requiring ORBit, just
yet :) If you want to try it with ORBit, just install the latest glib &
ORBit out of CVS.

There is a bug if you try to remove an applet from the panel - ideas
wanted.