[capplet] Big rework of the code to clean it up, fix bugs, etc.
The core of this rework is to separate in a clean way the code that
tracks the autostart files from the UI itself.

Now it'd be possible to make an autostart file inherit the data from
the autostart files with the same basename but in other directories.

This work also fixes a few bugs:
 + monitor autostart files
   http://bugzilla.gnome.org/show_bug.cgi?id=437204
 + make name field optional and autofill it instead
   http://bugzilla.gnome.org/show_bug.cgi?id=502393