[session-properties] set NoDisplay=true for the session properties dialog
It doesn't really fit into the new System Settings design nor does it really
make sense for a user to be able to turn off core desktop components like
power management or networking.  We'll likely redesign this to only
show applications and include it in some other place in the System Settings.