manager: remove client from query clients list on disconnect
When the session manager enters the query end session phase,
it temporarily adds all clients to the "query clients" list.

This list is used to track the progression of clients quitting.

If a client quits during the query end session phase instead of
waiting until the end session phase, then it will get destroyed
but not get removed from the query clients list.

This commit removes it from the query clients list in
this case. It does this by explicitly invoking the end session
response handler (in effect, the client responded to the end
session query by exiting).