gnome-session-check-accelerated: Set X property _GNOME_SESSION_ACCELERATED
It would make sense to run gnome-session-is-accelerated in
GDM, in parallel with the user is typing their password.  A
good way to pass that knowledge through to gnome-session
so it knows it can avoid repeating the same check is via
an X property.

So create a new tool "gnome-session-check-accelerated", which takes
the name of the old one (now renamed "gnome-session-check-accelerated-helper")
and stores that status as an X property.  gnome-session will then
itself run it again, but it should pick up the property and reuse that
value.

https://bugzilla.gnome.org/show_bug.cgi?id=641992