gsm-process-helper: Give useful error data
In order for engineers to be able to extract why a tool
like gnome-session-check-accelerated failed, we should print
something to ~/.xsession-errors.

gsm_process_helper() only returned the exit code, unless the process
was killed by a signal, in which case it returned -1.  However
the only consumer of the code never checked the exit code, just
success.

So fix this by having gsm_process_helper return a normal
gboolean/GError pair.  As part of this, clean up the code so that it
also handles the WIFSIGNALED and WIFSTOPPED cases (See "man 2 waitpid").

Also, the exit_child_simple() function was lame; we don't
need to call g_spawn_close_pid() on Unix, so don't do it.

https://bugzilla.gnome.org/show_bug.cgi?id=643880