manager: honor disable-log-out lockdown configuration
Right now we have a disable-log-out configuration key in the
org.gnome.desktop.lockdown schema that we superficially support in the
UI, but disregard when servicing the dbus session manager interface.

This commit makes gnome-session prohibit log out via its D-Bus
interface (and XSMP), if log out is locked down.

https://bugzilla.gnome.org/show_bug.cgi?id=645419