gsm: Use app-id when saving session instead of startup-id for .desktop
We try to use the app-id whenever possible for the .desktop filename
when saving an app since this means the saved app can be used as
required component on login. (Still fallback to startup-id if app-id
doesn't work because it's empty or filename already exists)

This is important because on login, the required component will be
loaded first, and if it's in the saved session with another name, the
saved version will get ignored: we first load all required components
from saved/autostart/apps, and then we load other apps in the saved
session (and the saved component, if using another name, will look like
a random app so will be loaded at that time).