shell: listen for 'Closed' signal
Right now we track when the shell end
session dialog is closed by tracking when the
user clicks cancel or (e.g.) logout.

The problem is, there are other cases besides
the user hitting a button on the dialog that leads
to the dialog getting closed.

This commit tracks the closed state of the dialog
via the closed signal instead of the button siganls.

Upon getting notification of the dialog closing,
we immediately stop processing updates to the
inhibitor list.

https://bugzilla.gnome.org/show_bug.cgi?id=645485