gsm: Disconnect all dbus clients when dbus is disconnected
When we receive the Disconnected signal from dbus on the session bus,
then we know we won't be able to talk to dbus clients anymore, so mark
them as disconnected. Also, do not restart them even if they're supposed
to be autorestarted, as we won't be able to track them.

At this point, the session is kind of hosed. We could possibly decide
to simply leave, but we don't do it in case it's a user session and
there's unsaved data.

This helps a lot in the case of the gdm greeter, see
https://bugzilla.gnome.org/show_bug.cgi?id=658481