Fix some problems in the systemd code
The stop and restart functions forgot to pass the manager
object down, and then tried to emit signals on a NULL
object. Also, we were unreffing a variant that was NULL
half the time. Don't do that.

Fixes the problem where a poweroff request in the presence
of multiple user sessions leads to a PolicyKit dialog that
comes back every time it is canceled.

https://bugzilla.gnome.org/show_bug.cgi?id=673619