Add an automount-inhibit flag
When connecting to virtual machines with usb-device redirection, such as
Spice enabled vms, automount may get in the way. Specifically if auto-usbredir
is enabled in the vm-viewer, then the usbredir code and the automount code
race for who gets to the device first.

If the automount code wins the race this is a problem, since usbredir causes
a device-disconnect (iow the usb mass storage driver sees an unplug), so in
the end usbredir always wins, and we end up with a non clean potentially
corrupt filesystem. Also see:
https://bugzilla.redhat.com/show_bug.cgi?id=812972

There for the need exists to be able to inhibit gnome-shell's automounting,
since all other inhibits run through gnome-session, it seems consistent to
store the inhibit flag for automounting in gnome-session too. This has the
added advantage that gnome-session already has all the necessary logic to
automatically uninhibit if a client goes away.

Another reason to do this through gnome-session is that all other inhibit
flags are accessible through GtkApplication which relies on gnome-session,
so storing this in gnome-session allows to add it to GtkApplication too.

Note that gnome-session does not do anything with the flag, other then
tracking it. The reason to still modify gnome-session is to reserve and
document the flag.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>

https://bugzilla.gnome.org/show_bug.cgi?id=678595