Make autostart overriding more flexible
This patch changes the interpretation of --autostart.

Previously, when specified on the commandline, no session file was
loaded at all. With this patch, we make it just override the locations
that we look for autostart files in, from the default
/etc/xdg/autostart
/usr/share/autostart
/usr/share/gnome/autostart

(we still consult those directories for fulfilling .session file
 requirements)

The goal of this is to enable gdm to use autostart in the login
session, so orca can be activated.

https://bugzilla.gnome.org/show_bug.cgi?id=663721