Reject new XSMP clients earlier in the shutdown path
When we are ending the session, a new client connecting and trying to
register causes things to get confused and long delays to happen.

Move GsmXsmpServer to be managed by GsmManager, which can now tell the
server at a much earlier time when we are shutting down. Now new clients
are rejected much earlier and the confusion and delays are gone.

https://bugzilla.gnome.org/show_bug.cgi?id=688209