main: disconnect shutdown_cb when manager is destroyed
shutdown_cb is supposed to be called when gnome-session is
running and suddenly the bus is ripped away from it.  Unfortunately,
it also gets run as part of the normal shutdown path, which it isn't
designed for.

This commit changes g_signal_connect_swapped to g_signal_connect_object
to ensure the handler is disconnected when the manager object goes away,
so the handler isn't called in the normal shutdown path.

https://bugzilla.gnome.org/show_bug.cgi?id=688992