manager: call complete_shutdown in better place
The code currently unconditionally calls complete_shutdown at the
bottom of the END_SESSION phase. This means it gets called at logout
time, and ConsoleKit users will get power off at logout time.

This commit moves it to be called in place of the superfluous calls
to attempt_stop/attempt_restart in gsm_manager_quit (which runs
at the proper time, during the EXIT phase, and has checks to exclude
the LOGOUT case)

https://bugzilla.gnome.org/show_bug.cgi?id=689721