Support automatic logout after a maximum period of idleness
This is very commonly requested in any situation with
shared resources: servers, labs, etc.

We simply start counting down once the session goes idle,
and display a warning notification when we get close to the
maximum allowed idle time.

The feature is controlled by two new org.gnome.SessionManager
settings: max-idle-time and max-idle-action. The time is specified
in minutes, and the action can be 'none', 'logout' or 'force-logout',
with 'none' being the default.

This commit adds a libnotify dependency.

http://bugzilla.gnome.org/show_bug.cgi?id=597030