session: Remove RequiredProviders support
This is only needed when we have interchangeable system components
to handle particular features, such as using a specific notification
daemon. We don't need that any more, as those should be provided by
RequiredComponents already, and not changeable.

If a non-GNOME desktop uses gnome-session, they will most likely have
their preferred provider for that feature.

https://bugzilla.gnome.org/show_bug.cgi?id=691663