presence: Handle really short idle delays
We wouldn't end up idle in the following scenario:
idle-delay = 10 secs
we've been idle for 7 secs
idle-delay = 5 secs
idle time is 7 secs, but our idle watch is for 5 seconds.

We handle this by manually checking for the idle time. In the case of a
race where the idle time might be triggered twice, it would not have
any adverse effects as set_session_idle() knows we're already idle.

https://bugzilla.gnome.org/show_bug.cgi?id=692395