gsm-manager: Don't abort when all clients exit
If all of our clients exit, remove_clients_for_connection()
calls end_phase() which in turn invokes gsm_manager_quit() that
checks the logout type.

But at this point we only have _LOGOUT_NONE, so we end up aborting.
It's fine to just call _quit() in this case, and this avoids us
dumping core.

This was happening to me when session startup failed for other
reasons.

https://bugzilla.gnome.org/show_bug.cgi?id=710480