manager: restart gnome-shell on X11 if it crashes
gnome-shell calls itself a DisplayServer for ordering reasons,
and because we share one file between wayland an X11.

gnome-session currently doesn't try to restart a DisplayServer if
it crashes, since the presumption is the old session will come down
with it.

This commit restarts "display server" applications on anything but
wayland, because they can't actually be a display server in other
cases.

https://bugzilla.gnome.org/show_bug.cgi?id=760332