Export GL/GLES renderer string over DBus
This is useful to avoid repeating the same code e.g. in the Details
settings panel.

Note that it's possible that both the display manager and the user
session run gnome-session, and that they both share the same X server.
In that case, we need to store the renderer string as an X property on
the root window (like the helper already does for other properties).

https://bugzilla.gnome.org/show_bug.cgi?id=686806