gnome-session: update activation environment at startup
The dbus daemon gets started before some environment variables
are set at start up.

This commit makes sure the dbus daemon is given a clean snapshot
of the environment at the time the session is started.

This is a near-term workaround for

https://github.com/alexlarsson/xdg-app/issues/147

and will probably go away when we figure out a story for bug 736660

https://bugzilla.gnome.org/show_bug.cgi?id=766176