main: Don't override IM variables
To give a user the option to use some other input method framework but
IBus, gnome-session should not override the QT_IM_MODULE and XMODIFIERS
environment variables if they are already set. The desired approach has
been implemented in gnome-settings-daemon:

https://git.gnome.org/browse/gnome-settings-daemon/commit/?id=21a16d25d981f3ce34b764dbc8005dbf8edfbe97

It would be consistent to change gnome-session in a similar way.