Do not report errors in absence of dbus-update-activation-environment
dbus-update-activation-environment was introduced with dbus 1.10.4
gnome-session by itself does not depend on any specific version of
dbus at all - but on dbus-glib-1 (DBUS_GLIB_REQUIRED=0.76)

The latest version of dbus-glib (ver 106) requires dbus-1 1.8: it is
thus a valid setup to run not have dbus-update-activation-environment
on the system. Warning the user about its absence is alerting users
for not reason.