gnome-session-inhibit: fix child cmd arguments
int execvp(const char *file, char *const argv[]);
The execv(), execvp(), and execvpe() functions provide an array of pointers
to null-terminated strings that represent the argument list available to the
new program. The first argument, by convention, should point to the filename
associated with the file being executed.

In the above paragraph "The first argument" is argv[0] (not the first
parameter of the execvp function).

https://bugzilla.gnome.org/show_bug.cgi?id=772825