gnome-session: update activation environment from C code, instead of shell script
dbus-update-activation-environment excepts certain environment
variables, that systemd won't.  We're going to want to eventually send
the environment to systemd, too, so we shouldn't make sure the same set
of variables get sent to both.

This commit takes the dbus-update-activation-environment call out of the
gnome-session shell script wrapper, and instead changes the C code to do
the export explicitly.

https://bugzilla.gnome.org/show_bug.cgi?id=736660