gsm-systemd.c: Restrict the login monitor to the "session" category
We only are interested in changes to sd_session_get_active(), so only
listen for changes in the "session" category. This reduces overhead
by a small amount, and also reduces the amount of /var/run/systemd
that needs to be read. (Bug 772537 is a failure because of a SELinux
policy preventint reading /var/run/systemd/machines - which we don't
need to monitor anyways.)