main: Start the manager only after we're sure we own the DBus name
This closes a race where the clients we spawn wouldn't be able to
reach us on the well known DBus name because they'd attempt to do so
before we own it.

By registering DBus APIs on the bus acquired callback but only
starting the GsmManager on name acquired we guarantee that clients
will be able to reach us.

https://bugzilla.gnome.org/show_bug.cgi?id=780405