save: make sure app state is written into desktop file
There are a number of important bits of app state written into
an applications desktop file that needs to be restored when
the session is saved.  For instance, the phase in which the client
should get started.

That state is currently stored on the GsmApp object, which is
inaccessible, from the client save function.

This commit adds the neccesary plumbing to route the GsmApp object
associated with a client to the client save function, and also
adds code to allow the app object to augment the client keyfile at
save time.

https://bugzilla.gnome.org/show_bug.cgi?id=790913