When there are no required apps (under systemd), let Setenv work
Setenv is a mechanism that some clients - notably mutter - use to set
environment variables for the use of things that are started later. It's
useful because it forwards changes to the systemd activation
environment, which is what units use.

Normally it's only available in early startup, but if starting things is
actually being handled by systemd and not gnome-session, we'll be
RUNNING without the session actually being "up". We were denying
mutter's Setenv calls, which set DISPLAY and WAYLAND_DISPLAY, quite
important environment variables.

Let's try saying that if there are no required apps (the systemd
situation), you can call Setenv while we're already running too.