Add support for new ConfirmedRebootMenu signal send by org.gnome.SessionManager.EndSessionDialog
In order to allow the user to choose advanced boot-options on systems where
the boot menu is not shown by default at boot, gnome-shell's endSessionDialog.js
may emit a new ConfirmedRebootMenu signal to indicate that gnome-session
should reboot the system in such a way that the boot-menu will be shown
next boot.

This implements the backend for the "Boot Options" button from:
https://wiki.gnome.org/Design/OS/BootOptions

Note that at the moment ConfirmedRebootMenu is just treated as an alias to
ConfirmedReboot, since there is no bootloader agnostic way yet to indicate
the menu should be shown.

For now the FIXME may be patched with a distro specific patch to implement
this for the distro's chosen bootloader. I will start a discussion on
systemd-devel to come up with a distro / bootloader agnostic interface
for this.