manager: Quit session manager after completing shutdown
Without this gnome-session-binary will simply continue to live when the
user requested a shutdown or reboot. In all cases (systemd, consolekit,
otherwise) it makes sense to simply quit g-s-m after completing the
shutdown.

In fact, we could just quit immediately in all cases, as this will
release the inhibitor FDs, but that wouldn't be quite as clean.