autostart-app: don't try to add child watch to dbus activated services
If a service is dbus activated, then gnome-session isn't starting it
directly and it makes no sense to try to create a child watch on it.

This commit ensures, child watches are only set up on apps that are
started directly by gnome-session, which avoids generating a critical
message in the log for each dbus activated app.

https://gitlab.gnome.org/GNOME/gnome-session/issues/14