data: Add systemd target for X11 specific services
These services should be a direct dependency of a X11 session, but
might be started or stopped as required on Wayland sessions. This
unit is pretty barebones, and it is expected that any service that
wants to be a dependency of it must set itself up as such.

Also, make gnome-session-x11.target require this new target, so
these X11 services are part of the X11 session initialization.