Delay XDG autostart applications on systemd
When launching on systemd we did not synchronize the phases after
the shell is loaded. Due to this, applications may start up before g-s-d
is started, which should not be the case.

Add a further one-way synchronization point so that gnome-session only
starts launching applications after gnome-session.target has been
reached.

Fixes: #35