data: Ensure shutdown target does not keep units loaded
The shutdown target has explicit conflicts on some user units, this
means that as long as it is active, it will keep those units loaded.

We could probably reverse the conflicts to prevent this from happening.
But, it makes sense to set StopWhenUnneeded=true anyway and doing so
also means that the corresponding units can be unloaded after logout.