gnome-session: avoid setting LC_ unless LANG and region disagree
At the moment, gnome-session explicitly sets the various LC_ variables
to the user configured region.

That's unnecessary, though.  If the LC_ variables are unset,
applications know to use LANG.  Furthermore, setting the LC_ variables
makes it so you can't override them with LANG anymore.

This commit makes sure the LC_ variables only get set when absolutely
necesary. That is, when they are different than LANG.

https://gitlab.gnome.org/GNOME/gnome-session/issues/37