save: Do not save apps in the RequiredComponents list
These are essential components of the session. They will be restored
anyway and should not be restored if e.g. the user switches sessions.

This actually caused trouble after the systemd move as gnome-session may
end up starting gnome-shell because it was auto-saved, while gnome-shell
is also started through systemd causing the session to fail.

See: #41