session-fill: Never look for required components in saved session
As required compoents are essential parts of the session, it does not
make sense to restore them automatically. Should e.g. the global
definition of the component change then we will want to pick up the new
definition rather than keeping the old one.

This addresses a migration issue to systemd startup where the shell may
be started from an auto-saved XDG desktop file.

The XDG desktop file will still be found as an autostart application
later in the filling process. However, this entry will be ignored as the
app is already registered in the store.

See: #41