data: Remove gnome-session-stable.timer and related units
The idea was that this is useful for gnome-shell to modify its OnFailure
behaviour. However, the approach has been causing issues with session
shutdown and the workaround is quite complicated.

To simplify matters, the logic has been moved to gnome-shell and further
improved. As such, the gnome-session-stable.timer has lost its purpose
and it is unlikely to be useful for other purposes in the future.

Remove it completely.

See gnome-shell!858 and #43