data: Permit checking gnome-session-stable during shutdown
gnome-shell tries to disable extensions after a failure, but only if the
session is crashing in the first two minutes. Unfortunately, this test
needs to be able to run at the same time as
gnome-session-shutdown.target startup, as both session shutdown and
extension disabling is part of the OnFailure action.

To make this work, allow testing for the existance of the
gnome-session-stable file in the runtime directory. This file is only
removed after gnome-session-shutdown.target is reached, which means
gnome-shell can use it in a ConditionPathExists as long as it also adds
Before=gnome-session-shutdown.target.

See #43