util: Add utility function to place a PID into a systemd scope
It makes sense to put children into their own systemd scope when
possible. This helper allows to do so, it is a direct copy of the
gnome-shell code.

In the gnome-session context we could short-cut e.g. the check whether
we are running on systemd, but keeping it as is does not add much
overhead either and means it is possible to unconditionally call the
functions without it resulting in unwanted warnings.