data: Configure session using a drop-in
During the GNOME 3.34 times when systemd support was still added, all
the session components pulled in themselves or were explicitly pulled by
GNOME session. However, doing this makes it really hard to create a
custom session or to modify the existing sessions in some way.

Change things so that we pull in all the dependencies for each session
type independently. This works using a drop-in for the
gnome-session@.target and all GNOME session users will need to provide
such a drop-in or pass --builtin to gnome-session.