tools: Add generator for session configuration drop-in
This simply generates Wants= entries for every RequiredComponent. We
generate Wants for the application ID directly, and also ones for the
wayland/x11 session with a -wayland/-x11 postfix added. This means that
applications can provide either a unit without any postfix, or a unit
with a session type specific postfix.

The session type specific postfix is solely intend to be used by
gnome-shell.