fill: Add AutostartWhitelist key to session file
gnome-session currently has an autostart directory override. The purpose
of this is to disable the usual autostart applications and only start
whatever is in the given autostart directory.

The nice thing about this is that it is possible (even for third
parties) to extend the autostart list. The bad part is that it is an odd
mechanism of partially disabling autostart. One that cannot be mapped
appropriately to e.g. a systemd generator.

To improve this, add a new AutostartWhitelist key. This key effecitvely
disables autostart (if left empty) and one can add a whitelist of
applications that should still be started (in the login screen, as that
is the only real user).

The disadvantage is that it means distributions need to e.g. add
spice-vdagent uconditionally to this list rather than installing it into
/usr/share/gdm/greeter/autostart. The advantage is that the key can be
appropriately resolved without passing any magic parameters or
environment variables over from GDM to the gnome-session. It also means
that quite a bit of code can be removed from gnome-session.