data: Delete the session leader fifo at shutdown time
To do so, update and rename the dbus-restart service to also delete the
file. Doing this means that we can check for the existence of the named
fifo (in most cases) to verify that a GNOME managed session is running.

It is not perfect, but good enough for some purposes.