systemd: Add autostart condition evaluator and monitoring service
This adds a new binary (installed to /usr/bin for use by systemd), which
can both evaluate autostart conditions and also monitor them. The
purpose of this binary is to be used as an ExecCondition expressions.

When used in this way, the service will only be running, when the
condition is true. This works by
 1. evaluating the condition and returning an appropriate value and
 2. registering the service for start/stop when the condition changes.

As such, the service should always be requested to start, but will only
actually be started as long as the condition is true. This creates an
equivalent behaviour as AutostartCondition= did for desktop files using
an ExecCondition= in a systemd service.

This is designed to be used by generators, in particular the XDG
autostart generator in systemd, see
  https://github.com/systemd/systemd/pull/15226