data: Leave to meson the duty of keeping the lists in sync
Don't leave to humans what machines can do, the gnome.session.desktop.in.in
is filled with an hard-to-maintain single-line semicolon separated list of
components, that then are repeated in the gnome systemd gnome.session.conf.
And these are meant to be kept in sync.

Meson can help us in this, so let's just keep a single list in meson and
use it to generate the various formats we need.

Using a map to keep the components so that this can be easily adapted and
possibly new sessions can be defined