Stop dbus-daemon instead of restarting it
On systems where dbus-daemon is a systemd user service, it is always
socket-activatable, so it will be restarted on-demand by dbus.socket.
On systems where it is not, neither StopUnit nor TryRestartUnit will
do anything. Either way, there doesn't seem much point in immediately
restarting it.

If the job contradicts jobs that are already queued, fail with a warning
instead of replacing those jobs.

Related to #74.

Signed-off-by: Simon McVittie <smcv@collabora.com>