presence: Do not auto-start the GNOME Shell screensaver service
As is, we are requesting the screensaver proxy to be started too early
in the login process. One issue that this is causing that login may hang
for a long time, other possible issues appear to be that the screensaver
service is not running at all in the end.

Note that gnome-settings-daemon will request the startup later on.

Thanks to Xiaoguang Wang <xwang@suse.com> for finding this issue and
suggesting a fix!

Closes: #88