data: Install GNOME on Wayland session for X11 preferred setups
gdm supports now setups where X11 is the preferred session, but wayland
can be used on demand.

However we're currently not proposing GNOME on Wayland session because
the preferred X11 GNOME session will be picked instead.

To avoid this, install (again) the gnome-wayland session, so that it
will be selectable when gdm is running under X11, but it also supports
wayland. Given its name, it will be filtered out in the case gdm is
running instead under wayland.