Drop consolekit backend and hard depend on systemd
Systemd has been the default session for a while and everyone not
using it is already heavily patching things and reimplements its
dbus apis (like logind).

This commit makes systemd a hard dependency, for both the
session and the journal, and drops the consolekit backend as
consolekit has been unmaintained for a good while now.