session: Remove x11 session targets
This is the first step towards deprecating the x11 session,
the gnome-xorg.desktop file is removed along with
the gnome-session-{x11, @} systemd targets.

But the x11 functionality is still there so it can be restored
by installing the files in the appropriate place again.

X11 has been receiving less and less testing. We have been
defaulting to the wayland session since 2016 and it's about
time we drop the x11 session completely.

Let's remove the desktop file and targets this cycle and maybe
carry on with removing rest of the x11 session code next cycle.

For removing the x11 code as well, see [1]

[1]: https://gitlab.gnome.org/GNOME/gnome-session/-/merge_requests/99