gnome-session wrapper: Avoid running as a subprocess
The gnome-settings-daemon 47.x test suite runs gnome-session as a
captive subprocess, terminating it after testing. If the gnome-session
wrapper script re-runs itself as a subprocess, then the subprocess will
eventually exec the actual gnome-session binary, but the parent will
continue to run.

This results in the g-s-d test suite eventually terminating the parent
but not the child, which breaks test suite assumptions for later tests
(which expect to get a new instance of gnome-session every time).

See https://gitlab.gnome.org/GNOME/gnome-settings-daemon/-/merge_requests/384

Tested-by: Alessandro Astone <alessandro.astone@canonical.com>
Signed-off-by: Simon McVittie <smcv@debian.org>