Drop the hardware acceleration check
We've always assumed that acceleration will work in Wayland
sessions, and Wayland has been the default for a while now.

This check is one of the remaining gtk3 dependencies in gnome-session

The list of known-bad hardware is also rather useless nowadays. Most of
the hardware listed there hails from ~2002, and some is old enough that
Mesa doesn't have drivers for it anymore.

Co-authored-by: Adrian Vovk <adrianvovk@gmail.com>