systemd: Drop blocking inhibitors before shutdown
Starting in systemd v257, systemd simplified the logic it uses for
inhibitors. It'll now enforce blocking inhibitors, even for processes
running as the same user that took them. So, gnome-session needs to make
sure it's not holding any blocking inhibitors when it's time to shutdown,
otherwise systemd will refuse to shutdown.

Fixes: https://gitlab.gnome.org/GNOME/gnome-session/-/issues/142