Drop session selector GUI
This is another component of gnome-session that depends on GTK3 and only
works on X11. This has been disabled by default since it was first
introduced in 2013 [1]. The only distro I could find that enables it is
RHEL, and in the upcoming RHEL 10 release there is no more Xorg session.

[1]: https://gitlab.gnome.org/GNOME/gnome-session/-/commit/6186a4a6f1fc1a34e188db4fe6d6e8035afea36d