Drop a legacy test
This is yet another legacy use of gtk3 in gnome-session. This test,
which is really more of a demo (it's not automatically run ever),
pops open a window and takes an inhibitor using the window's XID.

Nowadays, on Wayland, we don't have XIDs on windows so we always pass 0
as the XID. Since there's no need for a window, we can drop this demo's
dependency on GTK. But then at that point, this demo is the same as
`gnome-session-inhibit`. So, we can just drop it.