data: Ensure our app override is applied
gnome-desktop was ported to use the standardized systemd scope naming
convention [1], but gnome-session was never updated to apply its
override in the right place.

Also, apply the override to apps launched by Flatpak. That's a lot of
apps in a modern GNOME session, and in practice these apps don't get our
limits applied onto them.

[1]: https://gitlab.gnome.org/GNOME/gnome-desktop/-/commit/12367c11e8009bf7e2ff661f038e1e26a9642bc6